Revert "feat: real time auth updated"
This reverts commit 4f75b55c212c650342194f9e6c2bb21397e4f793.
